working on seek

+++ src/main/java/edu/brown/cs/am209hhe2lbenzonmsicat/sesh/Playlist.java
@@ -6,7 +6,6 @@ import java.util.Set;
 
 /**
  * Abstract playlist class.
- *
  * @author HE23
  */
 public abstract class Playlist {
@@ -50,7 +49,7 @@ public abstract class Playlist {
   //
   // public abstract void prevSong(String deviceId);
 
-  public abstract void seek(int position_ms, String deviceId);
+  public abstract void seek(long position_ms, String deviceId);
 
   /**
    * @param request
@@ -65,7 +64,6 @@ public abstract class Playlist {
 
   /**
    * This should only be used for testing!
-   *
    * @param spotifyId
    *          - spotify id
    * @param partyId
@@ -80,7 +78,6 @@ public abstract class Playlist {
 
   /**
    * Add playlist to database.
-   *
    * @param user
    *          - user
    * @return playlist
